requirements.txt
# Core libraries
numpy>=1.20.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
ultralytics>=8.0.0  # YOLOv8
matplotlib>=3.5.0
pandas>=1.3.0

# UI-related
PyQt5>=5.15.0
pyqtgraph>=0.12.0

# Utilities
PyYAML>=6.0
tqdm>=4.60.0
pillow>=8.0.0
